// enums/cnMaestro.enum.ts
export enum AdministrativeState {

    disable = "Disable",
    enable = "Enable",
}

export enum Device {

    switchGroup = "test1234",
    port1 = "Gi0/3",
    port2 = "Gi0/4",
    name = "Andrei-2028",
    portTypeHybrid = "Hybrid",
    portTypeTrunk = "Trunk",
    portTypeAccess = "Access",
    portPriority = "16",
    switchGroupNone = "None ",
    vlan = "20"
}

export enum StpMode {

    rstp = "RSTP",
    mstp = "MSTP",
    pvrst = "PVRST",
    rstpPriority = "8192",
    mstpPriority = "4096",
    pvrstPriority = "4096",
    regionNameMstp = "Reg1",
    vlan = "10",
    mstpVlans = "100"
}